smol/requirements.txt
smolagents
arize-phoenix
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-smolagents
